pyproject.toml
[project]
name = "pre-intership-healthcare-machine-learning-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
